Split up admin project poll
